requirements.txt
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.2
django-filter==24.3
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-silk==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.47
pillow==10.4.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.3
